Taiga: Make file name field wider on upload
This make several changes to the New Dataset/New Dataset Version dialog.
The code has been refactored to be more maintainable. The layout has
been updated to make the file name fields wider and more readable.
