+++ react_frontend/package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "aws-sdk": "^2.7.17",
     "bootstrap": "^3.3.7",
+    "classnames": "^2.5.1",
     "clipboard": "^1.6.0",
     "css-loader": "^5.2.4",
     "d3-array": "^1.2.1",
