+++ react_frontend/src/components/modals/UploadDialogs/UploadDialogs.scss
@@ -1,0 +1,76 @@
+@use "../../../styles/mixins";
+
+.Modal {
+  position: absolute;
+  margin: auto;
+  top: 30px;
+  left: 30px;
+  right: 30px;
+  bottom: 30px;
+  border-radius: 4px;
+  outline: none;
+  border: none;
+  padding: 0;
+  overflow: visible;
+  max-width: 1600px;
+
+  h2 {
+    margin-top: 11px;
+  }
+}
+
+.modalBody {
+  overflow: auto;
+  height: calc(100vh - 221px);
+  padding: 0 20px;
+  @include mixins.scroll-shadow;
+
+  input, textarea {
+    background-color: transparent;
+  }
+}
+
+.labelContainer {
+  label {
+    padding: 6px 0 5px 0;
+    margin: 0;
+    font-size: 18px;
+  }
+}
+
+.files {
+  margin-bottom: 0;
+
+  .labelContainer {
+    width: 100%;
+    position: sticky;
+    top: 0;
+    z-index: 1;
+    background-color: #fff;
+  }
+}
+
+
+.controls {
+  position: sticky;
+  top: 36px;
+  background-color: #fff;
+  z-index: 2;
+}
+
+.dropzone {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100px;
+  border: 2px rgb(102, 102, 102) dashed;
+  border-radius: 5px;
+  cursor: pointer;
+  background-color: #f5f5f5;
+}
+
+.gcsHelp {
+  margin-top: -8px;
+  margin-left: 8px;
+  margin-right: 12px;
+}
