+++ react_frontend/src/components/modals/UploadDialogs/UploadTable/UploadTable.scss
@@ -1,0 +1,27 @@
+.table {
+  thead {
+    position: sticky;
+    top: 185px;
+    z-index: 1;
+    border: none;
+    box-shadow: inset 0 -2px 0 #ccc;
+  }
+}
+
+.opaqueHeaderBackground {
+  margin-top: -35px;
+  position: sticky;
+  top: 185px;
+  min-height: 35px;
+  background-color: #fff;
+  z-index: 1;
+}
+
+.thRequestEncode {
+  width: 132px;
+}
+
+.thDelete {
+  width: 60px;
+}
+
