+++ taiga2/templates/index.html
@@ -356,17 +356,6 @@
   <body>
     <div id="root"></div>
 
-    <!-- TODO: Unfortunately, async attribute is giving an error on loading page. Chunking per page would be the best way to go -->
-    <!-- TODO: In production, use the minified libraries -->
-    <script src="//cdnjs.cloudflare.com/ajax/libs/react/15.4.1/react-with-addons.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/react/15.4.1/react-dom.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/react-router/2.7.0/ReactRouter.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/aws-sdk/2.7.17/aws-sdk.min.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/react-bootstrap/0.30.7/react-bootstrap.min.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/react-bootstrap-table/2.7.1/react-bootstrap-table.min.js"></script>
-    <script src="//netdna.bootstrapcdn.com/bootstrap/3.1.1/js/bootstrap.min.js"></script>
-
     <!-- Main -->
     <script>
       // Fetch the variables of the environment
